pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-steganography-studio"
version = "1.0.0"
description = "A Tkinter desktop application for hiding and revealing messages inside images."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Arsla" }
]
keywords = ["python", "steganography", "tkinter", "pillow", "desktop-app", "cryptography-learning"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Education",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Security :: Cryptography",
]
dependencies = [
  "Pillow>=10.0.0",
]

[project.scripts]
steganography-studio = "main:main"

[tool.setuptools]
packages = ["steganography"]
